backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
duckdb==0.9.2

# Authentication
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6

# LLM Integration
openai==1.3.7
anthropic==0.7.7

# FHIR
fhir.resources==7.1.0

# Data Processing
pandas==2.1.3
pyarrow==14.0.1
openpyxl==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
